Fix layout of main section and add dinamic sizing
The main layout changed to two column tracks and made the card section
a grid container with dinamic sizing.
Before this wasn't posible because this feature affected the size of the
announcements section (which had to be a fixed size).
Also, reorganized and some section dividers to the style.css file.